Move get_gcloud_storage_args to the util module
